🏗️ Organize code in domain folders and /plugins
