+++ src/files/name.rs
@@ -1,0 +1,8 @@
+use chrono::{Datelike, Local};
+
+pub fn name() -> String {
+    let today = Local::now().date_naive();
+    let day_of_month = today.day();
+    let file_name = format!("{:02}.md", day_of_month);
+    file_name.to_string()
+}
