+++ src/plugins/gdarquie_work/mod.rs
@@ -1,0 +1,1 @@
+pub mod work;
